+++ celescope/tools/plotly_plot.py
@@ -1,6 +1,8 @@
 import plotly
+import numpy as np
 import plotly.express as px
 import plotly.figure_factory as pf
+import plotly.graph_objects as go
 from collections import defaultdict
 from celescope.tools import utils
 
@@ -87,15 +89,7 @@ class Peak_plot(Plotly_plot):
         super().__init__(df)
         self.set_fig()
     
-    @staticmethod
-    def label(df):
-        if df["cell_called"] == True:
-            return "Cells"
-        else:
-            return "Non-cells"
-    
     def set_fig(self):
-        self._df['cell_called'] = self._df.apply(self.label, axis=1)
         self._df.sort_values("cell_called", ascending=False, inplace=True)
         
         self._fig = px.scatter(self._df, x="fragments", y="frac_peak", color="cell_called",color_discrete_sequence=["grey", "orange"]
@@ -113,6 +107,70 @@ class Peak_plot(Plotly_plot):
         )
         self._fig.update_traces(marker_size=3)
         
+
+class Frag_dis_plot(Plotly_plot):
+
+    def __init__(self, df):
+        super().__init__(df)
+        self.set_fig()
+    
+    @staticmethod
+    def get_plot_data(df):
+        singlecell_df = df
+        fragment_counts = singlecell_df['fragments'].values
+        cell_mask = (singlecell_df['cell_called']=='Cells')
+        noncell_mask = (singlecell_df['cell_called']=='Non-cells')
+
+        logbinmax = np.ceil(np.log10(fragment_counts.max()))
+        xbins = list(np.hstack([np.arange(100), np.logspace(np.log10(100), logbinmax, 350)]))
+        data_subplots = []
+        for name, mask, color in zip(["Non-cells", "{} Cells".format("")],
+                                     [noncell_mask, cell_mask],
+                                     ["blue", "orange"]):
+            # protect against really low depth samples, mixed up ref, severe under cell calling
+            if mask.sum() > 0 and logbinmax > np.log10(100):
+                counts, _ = np.histogram(fragment_counts[mask], xbins)
+                data_subplots.append({
+                    "name": name,
+                    "x": xbins,
+                    "y": list(counts),
+                    "type": "scatter",
+                    "connectgaps": True,
+                    "fill": "tozeroy",
+                    "line": {"color": color},
+                })
+    
+        return data_subplots
+    
+    def set_fig(self):
+        data_subplots = self.get_plot_data(self._df)
+        
+        self._fig = go.Figure()
+        self._fig.add_trace(
+            go.Scatter(
+                x=data_subplots[0]['x'], y=data_subplots[0]['y'],
+                fill='tozeroy', connectgaps=True, name='Non-cells'
+            )
+        ) # fill down to xaxis
+        self._fig.add_trace(
+            go.Scatter(
+                x=data_subplots[1]['x'], y=data_subplots[1]['y'],
+                fill='tozeroy', connectgaps=True, name='Cells'
+            )
+        ) # fill to trace0 y
+        
+        self._fig.update_layout(
+            width=470, height=313,
+            title={"text": "Fragment Distribution",
+                'y':0.98, 'x':0.5, 'xanchor': 'center', 'yanchor': 'top'},
+            xaxis={"type": "log", "color": "black", "gridcolor": "gainsboro", "linecolor": "black"},
+            yaxis={"type": "log", "color": "black", "gridcolor": "gainsboro", "linecolor": "black"},
+            xaxis_title = "Fragments per Barcode", yaxis_title = "Barcode", font=dict(size=12,color="Black"),
+            margin=dict(l=50, r=0, t=30, b=30),
+            plot_bgcolor="#FFFFFF",
+            legend=dict(title_text=''),
+        )
+
 
 class Umap_plot(Plotly_plot):